Add files via upload
A program that can change multiple files that are in a batch. I made this so I can rename my anime's episode to the same name so when you autoplay it, it will play in the order of each episode whereas before I needed to find the next episode in the folder.